src/requirements.txt
certifi==2020.4.5.1
Django==3.0.3
django-cors-headers==3.2.1
django-spa==0.3.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
wincertstore==0.2
asgiref==3.2.7
